Feat: 위도, 경도, 반경 기반으로 주변 경찰서 찾기 기능 완료

+++ src/main/java/com/example/buva/police/dto/PoliceFindReq.java
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.NotBlank;
 
 public record PoliceFindReq(@NotBlank(message = "위도를 입력해주세요") double latitude,
                             @NotBlank(message = "경도를 입력해주세요") double longitude,
-                            @NotBlank(message = "반경을 입력해주세요") double radius) {
+                            @NotBlank(message = "반경을 입력해주세요") Integer radius) {
 }
